requirements.txt
yfinance
pandas
numpy
supabase
python-dotenv
streamlit
plotly